// providers/keycloak/src/airflow/providers/keycloak/plugins/ui/openapi-gen/requests/types.gen.ts
// This file is auto-generated by @hey-api/openapi-ts

/**
 * HTTPException Model used for error response.
 */
export type HTTPExceptionResponse = {
  detail:
    | string
    | {
        [key: string]: unknown;
      };
};

export type HTTPValidationError = {
  detail?: Array<ValidationError>;
};

/**
 * Login serializer for post bodies.
 */
export type LoginBody = {
  username: string;
  password: string;
};

/**
 * Login serializer for responses.
 */
export type LoginResponse = {
  access_token: string;
};

export type UserRequest = {
  username: string;
  email: string;
  first_name: string;
  last_name: string;
  enabled?: boolean;
  email_verified?: boolean;
  role: string;
  password: string;
};

export type UserResponse = {
  userid: string;
  username: string;
  email: string;
  first_name: string;
  last_name: string;
  enabled: boolean;
  email_verified: boolean;
  roles: Array<string>;
  created_timestamp?: number | null;
};

export type UserUpdateRequest = {
  username?: string | null;
  email?: string | null;
  firstName?: string | null;
  lastName?: string | null;
  enabled?: boolean | null;
  emailVerified?: boolean | null;
  role?: string | null;
  password?: string | null;
};

export type ValidationError = {
  loc: Array<string | number>;
  msg: string;
  type: string;
};

export type CreateTokenData = {
  requestBody: LoginBody;
};

export type CreateTokenResponse = LoginResponse;

export type GetUsersResponse = Array<UserResponse>;

export type CreateUserData = {
  requestBody: UserRequest;
};

export type CreateUserResponse = UserResponse;

export type UpdateUserData = {
  requestBody: UserUpdateRequest;
  userid: string;
};

export type UpdateUserResponse = void;

export type DeleteUserData = {
  userid: string;
};

export type DeleteUserResponse = void;

export type $OpenApiTs = {
  "/auth/token": {
    post: {
      req: CreateTokenData;
      res: {
        /**
         * Successful Response
         */
        201: LoginResponse;
        /**
         * Bad Request
         */
        400: HTTPExceptionResponse;
        /**
         * Unauthorized
         */
        401: HTTPExceptionResponse;
        /**
         * Validation Error
         */
        422: HTTPValidationError;
      };
    };
  };
  "/auth/users": {
    get: {
      res: {
        /**
         * Successful Response
         */
        200: Array<UserResponse>;
        /**
         * Bad Request
         */
        400: HTTPExceptionResponse;
        /**
         * Unauthorized
         */
        401: HTTPExceptionResponse;
      };
    };
    post: {
      req: CreateUserData;
      res: {
        /**
         * Successful Response
         */
        201: UserResponse;
        /**
         * Bad Request
         */
        400: HTTPExceptionResponse;
        /**
         * Unauthorized
         */
        401: HTTPExceptionResponse;
        /**
         * Conflict
         */
        409: HTTPExceptionResponse;
        /**
         * Validation Error
         */
        422: HTTPValidationError;
      };
    };
  };
  "/auth/users/{userid}": {
    patch: {
      req: UpdateUserData;
      res: {
        /**
         * Successful Response
         */
        204: void;
        /**
         * Bad Request
         */
        400: HTTPExceptionResponse;
        /**
         * Unauthorized
         */
        401: HTTPExceptionResponse;
        /**
         * Conflict
         */
        409: HTTPExceptionResponse;
        /**
         * Validation Error
         */
        422: HTTPValidationError;
      };
    };
    delete: {
      req: DeleteUserData;
      res: {
        /**
         * Successful Response
         */
        204: void;
        /**
         * Bad Request
         */
        400: HTTPExceptionResponse;
        /**
         * Unauthorized
         */
        401: HTTPExceptionResponse;
        /**
         * Validation Error
         */
        422: HTTPValidationError;
      };
    };
  };
};